Fix truncated-prefix metrics on disconnect of statsrelay (#19)
Year old bug where statsrelay will disconnect in a write operation, but since it has no idea how framing works in statsd it decides to just write from where it left off, usually in the middle of a write.

This is dumb. 

This patch works around the problem by advancing the buffer forward until its after a newline (or end of buffer). This drops that metric, but its possible more than one metric was already dropped anyway due to the disconnect which isn't even accounted for.

We really need to change the backhaul to GRPC.

@lyft/observability