Fix whackdizzle logging (#63)
Iterating a va_list is destructive, so we need to use the va_copy macro
to grab a copy before looping over it with this asnprintf clone.

C99 is required. Also now caps the buffer at a maximum size and makes
the initial buffer bigger (since who even bothers counting a page of memory).
NOT PASSING TESTS